+++ site/src/utils/groups.ts
@@ -1,16 +1,5 @@
 import { Group } from "api/typesGenerated";
 
-export const everyOneGroup = (organizationId: string): Group => ({
-  id: organizationId,
-  name: "Everyone",
-  display_name: "",
-  organization_id: organizationId,
-  members: [],
-  avatar_url: "",
-  quota_allowance: 0,
-  source: "user",
-});
-
 /**
  * Returns true if the provided group is the 'Everyone' group.
  * The everyone group represents all the users in an organization
